requirements.txt
# incase requirements fail. use sudo pip install -r requirements.txt

Django==2.2.8
django-filter==2.0.0
django-password-reset==1.0
django-rest-framework-temporary-tokens==0.1.5
django-rest-passwordreset==0.9.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djoser==0.5.2
PyJWT==1.7.1
pytz==2019.2
sqlparse==0.3.0
gunicorn==19.6.0
Markdown==2.6.6
bleach==1.4.3
whitenoise==3.2
django-cors-headers==3.1.1
dry-rest-permissions==0.1.10
django-rest-swagger==2.2.0
psycopg2-binary
xlrd==1.2.0
django-crontab==0.7.1
africastalking==1.1.9
pyfcm==1.4.7
openpyxl==3.0.3